feat: 타입종류 추가 및 모달 내 타입 정의 분리
